Player name hover refactored
